Use a 32bit TTkey
Bench: 6716375

+++ src/types.h
@@ -4,7 +4,7 @@
 // define bitboard data type
 typedef uint64_t Bitboard;
 // define poskey data type
-typedef uint16_t TTKey;
+typedef uint32_t TTKey;
 // define poskey data type
 typedef uint64_t PosKey;
